+++ packages/webpack-model/src/modules-to-foam-tree.spec.ts
@@ -1,5 +1,5 @@
 import { Size } from '@statoscope/stats-extension-compressed/dist/generator';
-import stats from '../../../test/bundles/simple/stats-prod.json';
+import stats from '../../../test/bundles/v5/simple/stats-prod.json';
 import convert from './modules-to-foam-tree';
 import { NormalizedModule } from './normalize';