+++ test/fixtures/cli/validate/config-with-reporters.js
@@ -5,7 +5,7 @@ module.exports = {
     reporters: [['../../stats-validator/reporters/custom.js']],
     plugins: [['../../../../packages/stats-validator-plugin-webpack', 'webpack']],
     rules: {
-      'webpack/restricted-modules': ['error', [/\.\/src/]],
+      'webpack/restricted-modules': ['error', [/\/src\//]],
       'webpack/restricted-packages': ['error', ['foo']],
     },
   },